requirements.txt
Flask==2.2.2
python-dotenv==1.0.0
requests